Clean Up Chat Page
1. Create hooks for messages, message history, sendMessage, toggle privacy, etc.

2. Move reusable types to types folder
