+++ src/config/options.config.ts
@@ -1,0 +1,41 @@
+import { ISettingsList } from "@/types/settings"
+
+export const PII_TYPE_OPTIONS: ISettingsList[] = [
+    { entity: "ADDRESS", label: "Address" },
+    { entity: "AGE", label: "Age" },
+    // { entity: "ALL", label: "All" }, // Currently not needed --> perhaps use in future?
+    { entity: "AWS_ACCESS_KEY", label: "AWS Access Key" },
+    { entity: "AWS_SECRET_KEY", label: "AWS Secret Key" },
+    { entity: "BANK_ACCOUNT_NUMBER", label: "Bank Account Number" },
+    { entity: "BANK_ROUTING", label: "Bank Routing" },
+    { entity: "CA_HEALTH_NUMBER", label: "CA Health Number" },
+    { entity: "CA_SOCIAL_INSURANCE_NUMBER", label: "CA Social Insurance Number" },
+    { entity: "CREDIT_DEBIT_CVV", label: "Credit Debit CVV" },
+    { entity: "CREDIT_DEBIT_EXPIRY", label: "Credit Debit Expiry" },
+    { entity: "CREDIT_DEBIT_NUMBER", label: "Credit Debit Number" },
+    { entity: "DATE_TIME", label: "Date Time" },
+    { entity: "DRIVER_ID", label: "Driver ID" },
+    { entity: "EMAIL", label: "Email" },
+    { entity: "INTERNATIONAL_BANK_ACCOUNT_NUMBER", label: "International Bank Account Number" },
+    { entity: "IN_AADHAAR", label: "IN Aadhaar" },
+    { entity: "IN_NREGA", label: "IN NREGA" },
+    { entity: "IN_PERMANENT_ACCOUNT_NUMBER", label: "IN Permanent Account Number" },
+    { entity: "IN_VOTER_NUMBER", label: "IN Voter Number" },
+    { entity: "IP_ADDRESS", label: "IP Address" },
+    { entity: "LICENSE_PLATE", label: "License Plate" },
+    { entity: "MAC_ADDRESS", label: "MAC Address" },
+    { entity: "NAME", label: "Name" },
+    { entity: "PASSPORT_NUMBER", label: "Passport Number" },
+    { entity: "PASSWORD", label: "Password" },
+    { entity: "PHONE", label: "Phone" },
+    { entity: "PIN", label: "PIN" },
+    { entity: "SSN", label: "SSN" },
+    { entity: "SWIFT_CODE", label: "SWIFT Code" },
+    { entity: "UK_NATIONAL_HEALTH_SERVICE_NUMBER", label: "UK National Health Service Number" },
+    { entity: "UK_NATIONAL_INSURANCE_NUMBER", label: "UK National Insurance Number" },
+    { entity: "UK_UNIQUE_TAXPAYER_REFERENCE_NUMBER", label: "UK Unique Taxpayer Reference Number" },
+    { entity: "URL", label: "URL" },
+    { entity: "USERNAME", label: "Username" },
+    { entity: "US_INDIVIDUAL_TAX_IDENTIFICATION_NUMBER", label: "US Individual Tax Identification Number" },
+    { entity: "VEHICLE_IDENTIFICATION_NUMBER", label: "Vehicle Identification Number" }
+]
