+++ src/types/settings.d.ts
@@ -1,0 +1,4 @@
+export interface ISettingsList {
+    entity: string;
+    label: string;
+}
