+++ src/types/user.d.ts
@@ -1,0 +1,4 @@
+export interface IUser {
+    email: string | null;
+    username: string | null;
+}
